requirements.txt
Django==4.2.3

python-dotenv~=1.0.0
pytils~=0.4.1